requirements.txt
amqp==2.6.1
anyjson==0.3.3
asgiref==3.2.3
billiard==3.6.3.0
cairocffi==1.1.0
CairoSVG==2.4.2
celery==4.4.0
certifi==2020.4.5.1
cffi==1.14.0
cssselect2==0.3.0
defusedxml==0.6.0
Django==2.1
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.2.1
django-mathfilters==1.0.0
django-scheduler==0.9.3
django-widget-tweaks==1.4.6
gunicorn==20.0.4
html5lib==1.0.1
icalendar==4.0.4
importlib-metadata==1.7.0
kombu==4.6.11
Pillow==7.0.0
pycparser==2.20
PyMySQL==0.9.3
PyPDF2==1.26.0
Pyphen==0.9.5
python-dateutil==2.8.1
python-dotenv==0.13.0
python-http-client==3.2.6
pytz==2019.3
redis==3.5.3
reportlab==3.5.42
schedule==0.6.0
sendgrid==6.2.0
sentry-sdk==0.16.2
six==1.14.0
sqlparse==0.3.1
tinycss2==1.0.2
urllib3==1.25.9
vine==1.3.0
webencodings==0.5.1
xhtml2pdf==0.2.4
zipp==3.1.0
